Changed token position from long value to int value

+++ src/openscriptures/text/Token.java
@@ -20,7 +20,7 @@ public interface Token extends Comparable<Token> {
 	 */
 	public Work getWork();
 	
-	public long getPosition();
+	public int getPosition();
 
 	/**
 	 * Returns the type of this Token. One of, <tt>WORD</tt>, <tt>PUNCTUATION</tt>,
